Move HttpAPIHooks from proxy to tsapibackend
