+++ src/iocore/cache/CMakeLists.txt
@@ -48,8 +48,8 @@ target_include_directories(inkcache PRIVATE ${CMAKE_SOURCE_DIR}/lib)
 
 target_link_libraries(
   inkcache
-  PUBLIC ts::aio ts::hdrs ts::inkevent ts::tscore
-  PRIVATE ts::proxy fastlz ZLIB::ZLIB
+  PUBLIC ts::aio ts::hdrs ts::inkevent ts::tscore ts::tsapibackend
+  PRIVATE fastlz ZLIB::ZLIB
 )
 
 if(HAVE_LZMA_H)
